requirements.txt
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
colorama>=0.4.6
openai>=1.0.0
requests>=2.31.0
pytest>=7.4.0

